+++ CMakeLists.txt
@@ -623,14 +623,14 @@ set(NP_LIBS src/core/libraries/np/np_error.h
             src/core/libraries/np/np_manager.h
             src/core/libraries/np/np_matching2.cpp
             src/core/libraries/np/np_matching2.h
-            src/core/libraries/np/np_signaling.cpp
-            src/core/libraries/np/np_signaling.h
-            src/core/libraries/np/np_signaling_state.cpp
-            src/core/libraries/np/np_signaling_state.h
-            src/core/libraries/np/np_signaling_helpers.cpp
-            src/core/libraries/np/np_signaling_helpers.h
-            src/core/libraries/np/np_signaling_stubs.cpp
-            src/core/libraries/np/np_signaling_stubs.h
+            src/core/libraries/np/np_signaling/np_signaling.cpp
+            src/core/libraries/np/np_signaling/np_signaling.h
+            src/core/libraries/np/np_signaling/np_signaling_state.cpp
+            src/core/libraries/np/np_signaling/np_signaling_state.h
+            src/core/libraries/np/np_signaling/np_signaling_helpers.cpp
+            src/core/libraries/np/np_signaling/np_signaling_helpers.h
+            src/core/libraries/np/np_signaling/np_signaling_stubs.cpp
+            src/core/libraries/np/np_signaling/np_signaling_stubs.h
             src/core/libraries/np/np_trophy.cpp
             src/core/libraries/np/np_trophy.h
             src/core/libraries/np/np_tus.cpp
