+++ src/core/libraries/np/np_signaling/np_signaling.h
