+++ src/core/libraries/np/np_signaling/np_signaling_state.cpp
@@ -15,8 +15,8 @@
 #include "core/libraries/network/net.h"
 #include "core/libraries/network/net_util.h"
 #include "core/libraries/np/np_error.h"
-#include "core/libraries/np/np_signaling_state.h"
-#include "core/libraries/np/np_signaling_stubs.h"
+#include "core/libraries/np/np_signaling/np_signaling_state.h"
+#include "core/libraries/np/np_signaling/np_signaling_stubs.h"
 
 namespace Libraries::Np::NpSignaling {
