+++ src/core/libraries/np/np_signaling/np_signaling_stubs.h
